[mingw]  * do not perform sudo apt-get, unless needed  * do not perform sudo update-alternatives, unless needed

+++ mingw/build.sh
@@ -11,15 +11,25 @@ if test "$1" != "x86_64-w64-mingw32" && test "$1" != "i686-w64-mingw32"; then ec
 
 test -f ./Makefile && make clean || true
 
-sudo apt-get install mingw-w64-tools mingw-w64 libz-mingw-w64-dev nasm cmake
+if apt-get -s  install mingw-w64-tools mingw-w64 libz-mingw-w64-dev nasm cmake | grep -E -q "^(Inst|Conf|Remv)"; then
+    sudo apt-get install mingw-w64-tools mingw-w64 libz-mingw-w64-dev nasm cmake
+fi
 
 host=$1
 if test "$host" == "i686-w64-mingw32"; then
- sudo update-alternatives --set i686-w64-mingw32-g++ /usr/bin/i686-w64-mingw32-g++-posix
- sudo update-alternatives --set i686-w64-mingw32-gcc /usr/bin/i686-w64-mingw32-gcc-posix
+ if test `readlink -m \`which i686-w64-mingw32-g++\`|sed -e 's/\/.*\///'` != "i686-w64-mingw32-g++-posix"; then
+  sudo update-alternatives --set i686-w64-mingw32-g++ /usr/bin/i686-w64-mingw32-g++-posix
+ fi
+ if test `readlink -m \`which i686-w64-mingw32-gcc\`|sed -e 's/\/.*\///'` != "i686-w64-mingw32-gcc-posix"; then
+  sudo update-alternatives --set i686-w64-mingw32-gcc /usr/bin/i686-w64-mingw32-gcc-posix
+ fi
 else
- sudo update-alternatives --set x86_64-w64-mingw32-g++ /usr/bin/x86_64-w64-mingw32-g++-posix
- sudo update-alternatives --set x86_64-w64-mingw32-gcc /usr/bin/x86_64-w64-mingw32-gcc-posix
+ if test `readlink -m \`which x86_64-w64-mingw32-g++\`|sed -e 's/\/.*\///'` != "x86_64-w64-mingw32-g++-posix"; then
+  sudo update-alternatives --set x86_64-w64-mingw32-g++ /usr/bin/x86_64-w64-mingw32-g++-posix
+ fi
+ if test `readlink -m \`which x86_64-w64-mingw32-gcc\`|sed -e 's/\/.*\///'` != "x86_64-w64-mingw32-gcc-posix"; then
+  sudo update-alternatives --set x86_64-w64-mingw32-gcc /usr/bin/x86_64-w64-mingw32-gcc-posix
+ fi
 fi
 
 install=`pwd`/$host-install
